Order returned form documents by id
When dealing with pagination, if we do not order the results there is
the potential for results to move around in the order that the
database would return them, meaning results could be duplicated or
omitted when navigating through pages.

Order the results by their ID, as this is indexed by default and any
new records that are created when we are navigating through the pages
of results will be added at the end.