Use full precision timestamps in JSON responses
Currently we store timestamps with microsecond precision (the default
and maximum for PostgreSQL), but over the wire we only send milliseconds
(the default for Rails). This is fine for most purposes, but as we're
migrating data from forms-api to forms-admin at the moment it's helpful
to get the full precision so we can easily compare the contents of the
two databases without noise from truncated timestamps.

This commit uses the ActiveSupport configuration parameter to change the
`TimeWithZone#as_json` method to return all 6 digits of precision.