Remove next_page column from pages table
This column hasn't been used since before we migrated from grape (see
PR #116), when we migrated to grape we switched to using the
acts_as_list gem and the position column instead (see commit e8cc03).

Let's remove this column now. Note that we do still need the `next_page`
attribute in our JSON responses, as it used by consumers, but it is
calculated by a method based on the acts_as_list code, we don't
need it in our database.