Change response of /api/v1/forms/:id/make-live
forms-admin had been assuming that the response from POST-ing to the
make_live endpoint is a form object, but it's actually the made live
form document. This is less useful because it doesn't contain the state
(the attribute that is actually changed), and differs from the response
that we get from the archive endpoint, even though they're the
conceptually similar actions.

This commit changes the response to be the updated form object as
expected.